Slight refactoring
This makes the patient service the "master" copy of the current patient
data. Previously the search page held a copy of the data, now it always
gets it through the service. This will allow the data in the patient
service to be modified by the UI and those modifications to be used
everywhere.